requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-cors-headers==2.0.2
django-filter==2.0.0
django-modeltranslation==0.12.1
djangorestframework==3.9.0
djangorestframework-yaml==1.0.3
easy-thumbnails==2.5
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
olefile==0.46
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
PyYAML==3.12
requests==2.20.0
uritemplate==3.0.0
urllib3==1.23
uWSGI==2.0.17
